pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "trackly"
version = "0.1.6"
description = "Zero-overhead AI usage tracking for LangChain and LLM APIs"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [
    { name = "Udaykumar Dhokia", email = "udaykumardhokia@gmail.com" }
]
keywords = ["llm", "observability", "langchain", "openai", "anthropic", "cost-tracking"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries",
]

# SDK
dependencies = [
    "anthropic>=0.86.0",
    "google-genai>=1.68.0",
    "langchain-core>=0.2.0",
    "ollama>=0.6.1",
    "redis>=7.4.0",
    "requests>=2.28.0",
]

[project.optional-dependencies]
openai    = ["langchain-openai>=0.1.0"]
anthropic = ["langchain-anthropic>=0.1.0"]
gemini    = ["langchain-google-genai>=1.0.0"]
ollama    = ["langchain-ollama>=0.1.0"]
all       = [
    "langchain-openai>=0.1.0",
    "langchain-anthropic>=0.1.0",
    "langchain-google-genai>=1.0.0",
    "langchain-ollama>=0.1.0",
]
server    = [
    "reportlab>=4.4.10",
    "resend>=2.26.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.28.0",
    "pydantic-settings>=2.0.0",
    "slowapi>=0.1.9",
    "mangum>=0.17.0",
    "bcrypt>=4.0.0",
]
dev       = ["pytest>=7.0", "pytest-asyncio>=0.21"]

[project.urls]
Homepage      = "https://trytrackly.vercel.app"
Repository    = "https://github.com/udaykumar-dhokia/trackly"
"Bug Tracker" = "https://github.com/udaykumar-dhokia/trackly/issues"

[tool.setuptools.packages.find]
where   = ["."]
include = ["trackly*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
